% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_World.R
\name{hello_function}
\alias{hello_function}
\title{Hellow World Function}
\usage{
hello_function(first = TRUE)
}
\arguments{
\item{first}{What is the first program? Defauts to TRUE.}
}
\description{
This function allows you to know the first program.
}
\examples{
hello_function()
}
\keyword{program}
